+++ src/main/java/com/goorm/friendchise/global/auth/implement/jwt/TokenProvider.java
@@ -8,6 +8,8 @@ import io.jsonwebtoken.Jwts;
 import java.time.Duration;
 import java.util.Date;
 
+import static com.goorm.friendchise.global.auth.application.TokenExp.ACCESS_TOKEN_EXP;
+import static com.goorm.friendchise.global.auth.application.TokenExp.REFRESH_TOKEN_EXP;
 import static io.jsonwebtoken.Header.JWT_TYPE;
 import static io.jsonwebtoken.SignatureAlgorithm.HS256;
 import static javax.xml.crypto.dsig.SignatureProperties.TYPE;
@@ -18,7 +20,7 @@ import static javax.xml.crypto.dsig.SignatureProperties.TYPE;
 public class TokenProvider {
 	private final JwtProperties jwtProperties;
 
-	public String generateToken(String username, Duration expiredAt, String role) {
+	private String generateToken(String username, Duration expiredAt, String role) {
 		Date now = new Date();
 		Date expiry = new Date(now.getTime() + expiredAt.toMillis());
 		return makeToken(now, expiry, username, role);
@@ -37,5 +39,11 @@ public class TokenProvider {
 			.compact();
 	}
 
+	public String generateAccessToken(String username, String role) {
+		return generateToken(username, ACCESS_TOKEN_EXP.getExp(), role);
+	}
 
+	public String generateRefreshToken(String username, String role) {
+		return generateToken(username, REFRESH_TOKEN_EXP.getExp(), role);
+	}
 }
